Update UANodeSet.cs
In OPC UA V1.04, there were new access level bits defined. In the OPC UA information model, the attribute AccessLevel (byte) was not changed. Instead, a new optional attribute AccessLevelEx (Unsigned32) was added, whose lower 8 bits are identical to the the attribute AccessLevel.

In the XML NodeSet file, a different approach was chosen. Instead of adding another attribute, the existing attribute AccessLevel was changed from Byte to Unsigned32.
See https://opcfoundation.org/UA/schemas/1.04/UANodeSet.xsd.

+++ Stack/Opc.Ua.Core/Schema/UANodeSet.cs
@@ -1343,13 +1343,13 @@ namespace Opc.Ua.Export {
         
         /// <remarks/>
         [System.Xml.Serialization.XmlAttributeAttribute()]
-        [System.ComponentModel.DefaultValueAttribute(typeof(byte), "1")]
-        public byte UserAccessLevel {
+        [System.ComponentModel.DefaultValueAttribute(typeof(uint), "1")]
+        public uint UserAccessLevel {
             get {
                 return this.userAccessLevelField;
             }
             set {
-                this.userAccessLevelField = value;
+                this.userAccessLevelField = (byte)value;
             }
         }
